+++ examples/fetch_hacker_news.rs
@@ -3,11 +3,11 @@ use std::error::Error;
 use dom_query::Document;
 
 fn main() -> Result<(), Box<dyn Error>> {
-    let html: String = ureq::get("https://news.ycombinator.com/news")
+    let html = ureq::get("https://news.ycombinator.com/news")
         .call()?
         .into_string()?;
 
-    let document = Document::from(&html);
+    let document = Document::from(html.as_str());
 
     for news in document.select("tr.athing:has(a[href][id])").iter() {
         let link = news.select(".title span.titleline > a");