+++ src/property.rs
@@ -106,19 +106,13 @@ impl<'a> Selection<'a> {
     /// Gets the HTML contents of the first element in the set of matched
     /// elements. It includes the first matching element and its children nodes.
     pub fn html(&self) -> StrTendril {
-        match self.nodes().first() {
-            Some(node) => node.html(),
-            None => StrTendril::new(),
-        }
+        self.nodes.first().map_or_else(StrTendril::new, |node| node.html())
     }
 
     /// Gets the HTML contents of the first element in the set of matched
     /// elements. It includes only children nodes.
     pub fn inner_html(&self) -> StrTendril {
-        match self.nodes().first() {
-            Some(node) => node.inner_html(),
-            None => StrTendril::new(),
-        }
+        self.nodes.first().map_or_else(StrTendril::new, |node| node.inner_html())
     }
 
     /// Gets the HTML contents of the first element in the set of matched