+++ CHANGELOG.md
@@ -1,11 +1,15 @@
 # Changelog
 
 ## [0.4.2] - 2024-10-05
+
 ### Removed
 - Removed a `readability` examples and related dev-dependencies.
 
+### Added
+- Added `Selection::inner_html`, `Selection::try_html`, and `Selection::try_inner` methods
 
 ## [0.4.1] - 2024-10-04
+
 ### Fixed
 - Fix `Iterator::next` for `Matches`